thêm cart shopping va chinh sua 1 so loi

+++ database/migrations/2018_09_13_081550_create_orders_table.php
@@ -17,7 +17,7 @@ class CreateOrdersTable extends Migration
             $table->increments('id');
             $table->string('yourname');
             $table->integer('payment_id');//tra tien chua
-            $table->float('total');//Tong gia tien
+            $table->double('total');//Tong gia tien
             $table->string('email');
             $table->string('address');
             $table->string('phone');
